Fixes blurry text on horizontal orientation

+++ Classes/EasyTableView.m
@@ -23,7 +23,7 @@
 - (void) prepareForReuse
 {
     [super prepareForReuse];
-    
+
     UIView *content = [self viewWithTag:CELL_CONTENT_TAG];
     if ([content respondsToSelector:@selector(prepareForReuse)]) {
         [content performSelector:@selector(prepareForReuse)];
@@ -65,7 +65,7 @@
     if (self = [super initWithFrame:frame]) {
 		_numItems			= numCols;
 		_cellWidthOrHeight	= width;
-		
+
 		[self createTableWithOrientation:EasyTableViewOrientationHorizontal];
 	}
     return self;
@@ -76,7 +76,7 @@
     if (self = [super initWithFrame:frame]) {
 		_numItems			= numRows;
 		_cellWidthOrHeight	= height;
-		
+
 		[self createTableWithOrientation:EasyTableViewOrientationVertical];
     }
     return self;
@@ -86,28 +86,28 @@
 - (void)createTableWithOrientation:(EasyTableViewOrientation)orientation {
 	// Save the orientation so that the table view cell knows how to set itself up
 	_orientation = orientation;
-	
+
 	UITableView *tableView;
 	if (orientation == EasyTableViewOrientationHorizontal) {
-		int xOrigin	= (self.bounds.size.width - self.bounds.size.height)/2;
-		int yOrigin	= (self.bounds.size.height - self.bounds.size.width)/2;
+		CGFloat xOrigin	= (self.bounds.size.width - self.bounds.size.height)/2;
+		CGFloat yOrigin	= (self.bounds.size.height - self.bounds.size.width)/2;
 		tableView	= [[UITableView alloc] initWithFrame:CGRectMake(xOrigin, yOrigin, self.bounds.size.height, self.bounds.size.width)];
 	}
 	else
 		tableView	= [[UITableView alloc] initWithFrame:CGRectMake(0, 0, self.bounds.size.width, self.bounds.size.height)];
-	
+
 	tableView.tag				= TABLEVIEW_TAG;
 	tableView.delegate			= self;
 	tableView.dataSource		= self;
 	tableView.autoresizingMask	= UIViewAutoresizingFlexibleWidth | UIViewAutoresizingFlexibleHeight;
-	
+
 	// Rotate the tableView 90 degrees so that it is horizontal
 	if (orientation == EasyTableViewOrientationHorizontal)
 		tableView.transform	= CGAffineTransformMakeRotation(-M_PI/2);
-	
+
 	tableView.showsVerticalScrollIndicator	 = NO;
 	tableView.showsHorizontalScrollIndicator = NO;
-	
+
 	[self addSubview:tableView];
 }
 
@@ -123,7 +123,7 @@
 - (NSArray *)visibleViews {
 	NSArray *visibleCells = [self.tableView visibleCells];
 	NSMutableArray *visibleViews = [NSMutableArray arrayWithCapacity:[visibleCells count]];
-	
+
 	for (UIView *aView in visibleCells) {
 		[visibleViews addObject:[aView viewWithTag:CELL_CONTENT_TAG]];
 	}
@@ -133,20 +133,20 @@
 
 - (CGPoint)contentOffset {
 	CGPoint offset = self.tableView.contentOffset;
-	
+
 	if (_orientation == EasyTableViewOrientationHorizontal)
 		offset = CGPointMake(offset.y, offset.x);
-	
+
 	return offset;
 }
 
 
 - (CGSize)contentSize {
 	CGSize size = self.tableView.contentSize;
-	
+
 	if (_orientation == EasyTableViewOrientationHorizontal)
 		size = CGSizeMake(size.height, size.width);
-	
+
 	return size;
 }
 
@@ -161,7 +161,7 @@
 
 - (void)setContentOffset:(CGPoint)offset animated:(BOOL)animated {
 	CGPoint newOffset;
-	
+
 	if (_orientation == EasyTableViewOrientationHorizontal) {
 		newOffset = CGPointMake(offset.y, offset.x);
 	}
@@ -185,7 +185,7 @@
 - (void)selectCellAtIndexPath:(NSIndexPath *)indexPath animated:(BOOL)animated {
 	self.selectedIndexPath	= indexPath;
 	CGPoint defaultOffset	= CGPointMake(0, indexPath.row  *_cellWidthOrHeight);
-	
+
 	[self.tableView setContentOffset:defaultOffset animated:animated];
 }
 
@@ -193,16 +193,16 @@
 - (void)setSelectedIndexPath:(NSIndexPath *)indexPath {
 	if (![_selectedIndexPath isEqual:indexPath]) {
 		NSIndexPath *oldIndexPath = [_selectedIndexPath copy];
-		
+
 		_selectedIndexPath = indexPath;
-		
+
 		UITableViewCell *deselectedCell	= (UITableViewCell *)[self.tableView cellForRowAtIndexPath:oldIndexPath];
 		UITableViewCell *selectedCell	= (UITableViewCell *)[self.tableView cellForRowAtIndexPath:_selectedIndexPath];
-		
+
 		if ([delegate respondsToSelector:@selector(easyTableView:selectedView:atIndexPath:deselectedView:)]) {
 			UIView *selectedView = [selectedCell viewWithTag:CELL_CONTENT_TAG];
 			UIView *deselectedView = [deselectedCell viewWithTag:CELL_CONTENT_TAG];
-			
+
 			[delegate easyTableView:self
 					   selectedView:selectedView
 						atIndexPath:_selectedIndexPath
@@ -219,7 +219,7 @@
         UIView *headerView = [delegate easyTableView:self viewForHeaderInSection:section];
 		if (_orientation == EasyTableViewOrientationHorizontal)
 			return headerView.frame.size.width;
-		else 
+		else
 			return headerView.frame.size.height;
     }
     return 0.0;
@@ -230,7 +230,7 @@
         UIView *footerView = [delegate easyTableView:self viewForFooterInSection:section];
 		if (_orientation == EasyTableViewOrientationHorizontal)
 			return footerView.frame.size.width;
-		else 
+		else
 			return footerView.frame.size.height;
     }
     return 0.0;
@@ -241,9 +241,9 @@
 	// of the way UITableView resizes section views on orientation changes.
 	if (_orientation == EasyTableViewOrientationHorizontal)
 		sectionView.frame = CGRectMake(0, 0, sectionView.frame.size.width, self.frame.size.height);
-	
+
 	UIView *rotatedView = [[UIView alloc] initWithFrame:sectionView.frame];
-	
+
 	if (_orientation == EasyTableViewOrientationHorizontal) {
 		rotatedView.transform = CGAffineTransformMakeRotation(M_PI/2);
 		sectionView.autoresizingMask = UIViewAutoresizingFlexibleTopMargin;
@@ -272,7 +272,7 @@
 }
 
 -(NSInteger)numberOfSectionsInTableView:(UITableView *)tableView {
-    
+
     if ([delegate respondsToSelector:@selector(numberOfSectionsInEasyTableView:)]) {
         return [delegate numberOfSectionsInEasyTableView:self];
     }
@@ -289,12 +289,12 @@
 
 - (NSIndexPath *)indexPathForView:(UIView *)view {
 	NSArray *visibleCells = [self.tableView visibleCells];
-	
+
 	__block NSIndexPath *indexPath = [NSIndexPath indexPathForRow:0 inSection:0];
-	
+
 	[visibleCells enumerateObjectsUsingBlock:^(id obj, NSUInteger idx, BOOL *stop) {
 		UITableViewCell *cell = obj;
-        
+
 		if ([cell viewWithTag:CELL_CONTENT_TAG] == view) {
             indexPath = [self.tableView indexPathForCell:cell];
 			*stop = YES;
@@ -306,7 +306,7 @@
 - (CGPoint)offsetForView:(UIView *)view {
 	// Get the location of the cell
 	CGPoint cellOrigin = [view convertPoint:view.frame.origin toView:self];
-	
+
 	// No need to compensate for orientation since all values are already adjusted for orientation
 	return cellOrigin;
 }
@@ -340,13 +340,13 @@
 - (void)scrollViewDidScroll:(UIScrollView *)scrollView {
 	if ([delegate respondsToSelector:@selector(easyTableView:scrolledToOffset:)])
 		[delegate easyTableView:self scrolledToOffset:self.contentOffset];
-	
+
 	CGFloat amountScrolled	= self.contentOffset.x;
 	CGFloat maxScrollAmount = [self contentSize].width - self.bounds.size.width;
-	
+
 	if (amountScrolled > maxScrollAmount) amountScrolled = maxScrollAmount;
 	if (amountScrolled < 0) amountScrolled = 0;
-	
+
 	if ([delegate respondsToSelector:@selector(easyTableView:scrolledToFraction:)])
 		[delegate easyTableView:self scrolledToFraction:amountScrolled/maxScrollAmount];
 }
@@ -367,46 +367,46 @@
 
 - (UITableViewCell *)tableView:(UITableView *)aTableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {
     static NSString *CellIdentifier = @"EasyTableViewCell";
-    
+
     UITableViewCell *cell = [aTableView dequeueReusableCellWithIdentifier:CellIdentifier];
     if (cell == nil) {
         cell = [[EasyTableViewCell alloc] initWithStyle:UITableViewCellStyleDefault reuseIdentifier:CellIdentifier];
 
 		cell.backgroundColor = self.cellBackgroundColor;
 		[self setCell:cell boundsForOrientation:_orientation];
-		
+
 		cell.contentView.frame = cell.bounds;
 		cell.selectionStyle = UITableViewCellSelectionStyleNone;
-		
+
 		// Add a view to the cell's content view that is rotated to compensate for the table view rotation
 		CGRect viewRect;
 		if (_orientation == EasyTableViewOrientationHorizontal)
 			viewRect = CGRectMake(0, 0, cell.bounds.size.height, cell.bounds.size.width);
 		else
 			viewRect = CGRectMake(0, 0, cell.bounds.size.width, cell.bounds.size.height);
-		
+
 		UIView *rotatedView				= [[UIView alloc] initWithFrame:viewRect];
 		rotatedView.tag					= ROTATED_CELL_VIEW_TAG;
 		rotatedView.center				= cell.contentView.center;
 		rotatedView.backgroundColor		= self.cellBackgroundColor;
-		
+
 		if (_orientation == EasyTableViewOrientationHorizontal) {
 			rotatedView.autoresizingMask = UIViewAutoresizingFlexibleHeight;
 			rotatedView.transform = CGAffineTransformMakeRotation(M_PI/2);
 		}
-		else 
+		else
 			rotatedView.autoresizingMask = UIViewAutoresizingFlexibleHeight | UIViewAutoresizingFlexibleWidth;
-		
+
 		// We want to make sure any expanded content is not visible when the cell is deselected
 		rotatedView.clipsToBounds = YES;
-		
+
 		// Prepare and add the custom subviews
 		[self prepareRotatedView:rotatedView];
-		
+
 		[cell.contentView addSubview:rotatedView];
 	}
 	[self setCell:cell boundsForOrientation:_orientation];
-	
+
 	[self setDataForRotatedView:[cell.contentView viewWithTag:ROTATED_CELL_VIEW_TAG] forIndexPath:indexPath];
     return cell;
 }
@@ -414,15 +414,15 @@
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
 	NSUInteger numOfItems = _numItems;
-	
+
 	if ([delegate respondsToSelector:@selector(numberOfCellsForEasyTableView:inSection:)]) {
 		numOfItems = [delegate numberOfCellsForEasyTableView:self inSection:section];
-		
+
 		// Animate any changes in the number of items
 		[tableView beginUpdates];
 		[tableView endUpdates];
 	}
-	
+
     return numOfItems;
 }
 
@@ -431,11 +431,11 @@
 
 - (void)prepareRotatedView:(UIView *)rotatedView {
 	UIView *content = [delegate easyTableView:self viewForRect:rotatedView.bounds];
-	
+
 	// Add a default view if none is provided
 	if (content == nil)
 		content = [[UIView alloc] initWithFrame:rotatedView.bounds];
-	
+
 	content.autoresizingMask = UIViewAutoresizingFlexibleHeight | UIViewAutoresizingFlexibleWidth;
 	content.tag = CELL_CONTENT_TAG;
 	[rotatedView addSubview:content];
@@ -444,7 +444,7 @@
 
 - (void)setDataForRotatedView:(UIView *)rotatedView forIndexPath:(NSIndexPath *)indexPath {
 	UIView *content = [rotatedView viewWithTag:CELL_CONTENT_TAG];
-	
+
    [delegate easyTableView:self setDataForView:content forIndexPath:indexPath];
 }
